Fixing Code42's indecisiveness (#3)
With crashplan 6.8.8 and under, the log was stored in `history.log.0`. In 7.0.0, they changed the log to `history.log`. In 7.2 however, they changed it back to the original!
This code simple checks for the existence of both and figures out which one is newer. It is terrible code and there is probably a cleaner way to do it. However, I believe it does work.